organize actions, even better docstrings
