+++ cibo/actions/login.py
@@ -1,0 +1,20 @@
+"""Login Action"""
+
+from typing import List
+
+from cibo.actions import Action
+from cibo.models.client import Client
+
+
+class Login(Action):
+    """Log in to an existing player on the server."""
+
+    def required_args(self) -> List[str]:
+        """Descriptions of the args required for the action."""
+
+        return ["name", "password"]
+
+    def process(self, client: Client, args: List[str]):
+        """Process the logic for the action."""
+
+        _ = client, args
