+++ cibo/actions/look.py
@@ -1,0 +1,20 @@
+"""Look Action"""
+
+from typing import List
+
+from cibo.actions import Action
+from cibo.models.client import Client
+
+
+class Look(Action):
+    """Returns information about the room or object targeted."""
+
+    def required_args(self) -> List[str]:
+        """Descriptions of the args required for the action."""
+
+        return []
+
+    def process(self, client: Client, args: List[str]):
+        """Process the logic for the action."""
+
+        _ = client, args
