+++ cibo/actions/move.py
@@ -1,0 +1,20 @@
+"""Move Action"""
+
+from typing import List
+
+from cibo.actions import Action
+from cibo.models.client import Client
+
+
+class Move(Action):
+    """Moves a client between available rooms."""
+
+    def required_args(self) -> List[str]:
+        """Descriptions of the args required for the action."""
+
+        return []
+
+    def process(self, client: Client, args: List[str]):
+        """Process the logic for the action."""
+
+        _ = client, args
