+++ cibo/models/__init__.py
@@ -1,0 +1,5 @@
+"""Models and schemas that are used to represent and validate a database object.
+
+Also contains dataclasses that are used throughout server logic, but aren't required
+to be persistent.
+"""
