+++ cibo/models/item.py
@@ -1,4 +1,6 @@
-"""Item model"""
+"""An item is an in-game piece of inventory, that can be picked up, carried, and
+often used by a player.
+"""
 
 from dataclasses import dataclass
