+++ cibo/models/player.py
@@ -1,10 +1,12 @@
-"""Player model"""
+"""A Player represents a playable character, which a client logs into and assumes in
+order to interact with the world. Player information is persistent, allowing the client
+to resume their adventure where they left offE.
+"""
 
 
 from marshmallow import Schema, fields, validate
 from peewee import AutoField, CharField, TextField
 
-from cibo.helpers.models import field_is_alphanumeric
 from cibo.models.database import DatabaseModel
 
 
@@ -21,6 +23,6 @@ class PlayerSchema(Schema):
 
     id_ = fields.Int()
     name = fields.Str(
-        validate=[validate.Length(min=3, max=15), field_is_alphanumeric()]
+        validate=[validate.Length(min=3, max=15), validate.Regexp("^[a-zA-Z0-9_]*$")]
     )
     password = fields.Str(validate=validate.Length(min=8))
