+++ cibo/output.py
@@ -5,6 +5,8 @@ from typing import List
 
 from cibo.models.client import Client
 
+# TODO: figure out what to do with this
+
 
 class Output:
     """Responsible for constructing messages that are sent to a client."""
